+++ lib/generators/airbrake_initializer.rb.erb
@@ -33,15 +33,7 @@ Airbrake.configure do |c|
   # By default, Airbrake Ruby outputs to STDOUT. In Rails apps it makes sense to
   # use the Rails' logger.
   # https://github.com/airbrake/airbrake-ruby#logger
-  c.logger =
-    if ENV['RAILS_LOG_TO_STDOUT'].present?
-      Logger.new(STDOUT, level: Rails.logger.level)
-    else
-      Logger.new(
-        Rails.root.join('log', 'airbrake.log'),
-        level: Rails.logger.level
-      )
-    end
+  c.logger = Airbrake::Rails.logger
 
   # Configures the environment the application is running in. Helps the Airbrake
   # dashboard to distinguish between exceptions occurring in different
